Move address cache to separate file
